requirements.txt
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.3.1
torch==1.9.0
torchvision==0.10.0
scikit-learn==0.24.2
